Rework article listeners
