+++ app/src/main/java/com/capyreader/app/ui/articles/detail/ArticleView.kt
@@ -144,7 +144,5 @@ fun ArticleView(
 
     ArticleTemplateColorListener(webView = webViewState.webView, templateColors = templateColors)
 
-    ArticleTextSizeListener(webView = webViewState.webView)
-
-    ArticleFontFamilyListener(webView = webViewState.webView)
+    ArticleStyleListener(webView = webViewState.webView)
 }
